fix(desktop, host-internal): 修复 PR Tab 导入 host-internal 主入口导致 renderer 编译失败
