+++ apps/desktop/scripts/check-renderer-host-internal-imports.mjs
@@ -17,6 +17,7 @@ const RENDERER_SAFE_HOST_INTERNAL_SUBPATHS = new Set([
   'openai-models',
   'skill-paths',
   'github-pull-request-url',
+  'github-pull-request-checks-pages',
 ]);
 
 const RENDERER_SCAN_ROOTS = ['components', 'hooks', 'lib', 'App.tsx'];
